+++ resources/views/accounting/components/page-header.blade.php
@@ -17,21 +17,19 @@
          @if($createRoute)
             @if($createPermission)
                 @can($createPermission)
-                    <a href="{{ $createRoute }}"
+                    <a href="{{ $createRoute }}" title="{{ $createLabel }}"
                         class="inline-flex items-center px-4 py-2 bg-blue-950 border border-transparent rounded-md font-semibold text-xs text-white uppercase tracking-widest hover:bg-green-950 focus:bg-green-800 active:bg-green-800 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2 transition ease-in-out duration-150">
-                        <svg  class="size-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                        <svg class="size-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                             <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 4v16m8-8H4" />
                         </svg>
-                        <span class="hidden md:inline-block">{{ $createLabel }}</span>
                     </a>
                 @endcan
             @else
-                <a href="{{ $createRoute }}"
+                <a href="{{ $createRoute }}" title="{{ $createLabel }}"
                     class="inline-flex items-center px-4 py-2 bg-blue-950 border border-transparent rounded-md font-semibold text-xs text-white uppercase tracking-widest hover:bg-green-950 focus:bg-green-800 active:bg-green-800 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2 transition ease-in-out duration-150">
-                    <svg  class="size-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                    <svg class="size-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                         <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 4v16m8-8H4" />
                     </svg>
-                    <span class="hidden md:inline-block">{{ $createLabel }}</span>
                 </a>
             @endif
         @endif